# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Baahar API
  version: 0.1.0
  summary: Source-backed public city events
  description: >-
    The contract for Baahar's public discovery feed and protected source
    operations. Public event facts come from verified source versions; missing
    optional values remain null rather than being inferred.
servers:
  - url: http://127.0.0.1:8080
    description: Local API
...
paths:
  /v1/cities:
    get:
      operationId: listCities
      tags: [public]
      summary: List enabled cities
      responses:
        "200":
          description: Enabled city choices
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/City"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    operatorToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
  headers:
    ETag:
      description: Entity tag for conditional requests
      schema:
        type: string
...
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    CitySlug:
      type: string
      minLength: 1
      maxLength: 80
      pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
...
    City:
      type: object
      additionalProperties: false
      required: [slug, name, timezone, accent]
      properties:
        slug:
          $ref: "#/components/schemas/CitySlug"
        name:
          type: string
          minLength: 1
          maxLength: 100
        timezone:
          type: string
          minLength: 1
          maxLength: 100
        accent:
          type: string
          minLength: 1
          maxLength: 40
...
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          pattern: "^[a-z0-9_]+$"